extract latin1 to textutil
